Added 3 of 5 functions
In this patch, the create item, list all items, and exit functions are working properly. For the next patch, code will be modified so that user's do not get to choose the products number as this can cause issues when a user accidentally (or intentionally) types the same product number for two different products.